showcase-launchpad/storage/requirements.txt
fastapi==0.116.0
uvicorn==0.35.0
python-dotenv==1.1.1
slowapi==0.1.9
requests==2.32.4
PyJWT==2.10.1
pydantic==2.11.7
aiofiles==24.1.0
python-multipart==0.0.20
